IOS-6767: Rename and refactor
